+++ sudoku-solver/.gitignore
@@ -22,5 +22,7 @@ npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
 
-#personal notes
+#personal notes and sketches
 notes.txt
+sketches.png
+sketches.psd
